Move registrar directo logic to concern, touch up translations
