+++ app/models/invoice.rb
@@ -2,6 +2,7 @@ class Invoice < ApplicationRecord
   include Versions
   include Concerns::Invoice::Cancellable
   include Concerns::Invoice::Payable
+  include Concerns::Invoice::BookKeeping
 
   belongs_to :buyer, class_name: 'Registrar'
   has_one  :account_activity
@@ -100,24 +101,6 @@ class Invoice < ApplicationRecord
   def to_e_invoice
     generator = Invoice::EInvoiceGenerator.new(self)
     generator.generate
-  end
-
-  def as_directo_json
-    inv = ActiveSupport::JSON.decode(ActiveSupport::JSON.encode(self))
-    inv['customer_code'] = buyer.accounting_customer_code
-    inv['issue_date'] = issue_date.strftime('%Y-%m-%d')
-    inv['transaction_date'] = account_activity.bank_transaction&.paid_at&.strftime('%Y-%m-%d')
-    inv['language'] = buyer.language
-    inv['invoice_lines'] = compose_directo_product
-
-    inv
-  end
-
-  def compose_directo_product
-    [{ 'product_id': Setting.directo_receipt_product_name, 'description': order,
-       'quantity': 1, 'price': ActionController::Base.helpers.number_with_precision(
-         subtotal, precision: 2, separator: '.'
-       ) }].as_json
   end
 
   def do_not_send_e_invoice?
